refactor: Move DeltaManagerProxyBase
